Create portfolio
This code uses the Flask web framework to create a web application with routes for each of the pages you specified. Each route is associated with a function that returns the corresponding template. The templates (home.html, about.html, etc.) would need to be created separately.

Note that this is just a skeleton code and there is a lot more to add like

styling
Database integration
Adding functionalities to the pages
Error handling
Security measures
Deployment